requirements.txt
numpy
ipython
ipywidgets
kagglehub
plotly
torch
torchmetrics
torchvision
jupyter
pandas
torch_tb_profiler
tensorflow
gcsfs
optuna
plotly
kaleido
google-cloud-storage
matplotlib
streamlit
google-cloud-aiplatform
protobuf>=3.20,<5
pillow
streamlit
google-cloud-aiplatform
Pillow
torch-model-archiver
ts